+++ src/controller/controller.h
@@ -5,7 +5,7 @@ class View;
 class Controller
 {
 public:
-    static Controller *getInstance(unsigned int width = 0, unsigned int height = 0, const char *name = "");
+    static Controller *getInstance(unsigned short width = 0, unsigned short height = 0, const char *name = "");
     static void destroyInstance();
 
     void run();
@@ -15,7 +15,7 @@ public:
 
 private:
     static Controller *instance;
-    Controller(unsigned int width, unsigned int height, const char *name);
+    Controller(unsigned short width, unsigned short height, const char *name);
     ~Controller();
 
     View *view = nullptr;
