+++ src/view/view.h
@@ -5,7 +5,7 @@ class Controller;
 class View
 {
 public:
-    static View *getInstance(unsigned int width = 0, unsigned int height = 0, const char *name = "");
+    static View *getInstance(unsigned short width = 0, unsigned short height = 0, const char *name = "");
     static void destroyInstance();
 
     void display(unsigned int ebo, Controller & controller);
@@ -22,6 +22,6 @@ private:
     unsigned int width, height;
     bool initialized = false;
 
-    View(unsigned int width, unsigned int height, const char *name);
+    View(unsigned short width, unsigned short height, const char *name);
     ~View();
 };
